test: remove subfolder structure ang. dist. tests
