+++ tests/physics/helicity_formalism/angular_distributions/.gitignore
@@ -1,0 +1,1 @@
+!*.yml
