--- tests/physics/helicity_formalism/angular_distributions/d1_to_d0_pi_pi_decay/__init__.py
@@ -1,1 +1,0 @@
-"""Required to set mypy options for the tests folder."""
