--- tests/physics/helicity_formalism/angular_distributions/epem_to_dm_d0_pip_decay/__init__.py
@@ -1,1 +1,0 @@
-"""Required to set mypy options for the tests folder."""
